added cache for user-trips, fixed routes, imports  and renamed files
