+++ app.js
@@ -13,7 +13,7 @@ const { errorMiddleware } = require('./src/middleware/error');
 const communityRouter = require('./src/routes/communityRouter');
 const authRouter = require('./src/routes/auth_routes');
 const tripRouter = require('./src/routes/tripRouter');
-const userRouter = require('./src/routes/user_routes');
+const userRouter = require('./src/routes/userRouter');
 
 //Additional middlewares
 app.use(express.json())
