+++ src/utils/cache.js
@@ -1,0 +1,9 @@
+
+function removeElementFromSet(set, elementToRemove) {
+    if (set.has(elementToRemove)) {
+        set.delete(elementToRemove);
+    }
+    return set;
+}
+
+module.exports = { removeElementFromSet }
